refactor: 피드백 사항 적용.
