+++ src/test/java/nextstep/subway/favorite/FavoriteAcceptanceTest.java
@@ -12,12 +12,10 @@ import static nextstep.subway.member.MemberFixture.회원_생성을_요청;
 import static nextstep.subway.station.StationFixture.지하철역_등록되어_있음;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.station.dto.StationResponse;
@@ -25,7 +23,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 @DisplayName("즐겨찾기 관련 기능")
 public class FavoriteAcceptanceTest extends AcceptanceTest {
@@ -35,41 +32,6 @@ public class FavoriteAcceptanceTest extends AcceptanceTest {
 
     private String accessToken;
     private String newAccessToken;
-
-    public static ExtractableResponse<Response> 즐겨찾기_생성(final String accessToken, Long source,
-        Long target) {
-        FavoriteRequest favoriteRequest = new FavoriteRequest(source, target);
-
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(accessToken)
-            .body(favoriteRequest)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().post("/favorites")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 즐겨찾기_목록_조회(final String accessToken) {
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(accessToken)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/favorites")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(final String accessToken,
-        final String uri) {
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(accessToken)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().delete(uri)
-            .then().log().all()
-            .extract();
-    }
 
     @BeforeEach
     public void setUp() {
@@ -106,19 +68,19 @@ public class FavoriteAcceptanceTest extends AcceptanceTest {
     @Test
     void manageFavorite() {
         //when
-        ExtractableResponse<Response> createResponse = 즐겨찾기_생성(accessToken, 강남역.getId(),
+        ExtractableResponse<Response> createResponse = FavoriteFixture.즐겨찾기_생성(accessToken, 강남역.getId(),
             정자역.getId());
         //then
         즐겨찾기_생성됨(createResponse);
 
         //when
-        List<FavoriteResponse> getFavoriteResponses = 즐겨찾기_목록_조회(accessToken).jsonPath()
+        List<FavoriteResponse> getFavoriteResponses = FavoriteFixture.즐겨찾기_목록_조회(accessToken).jsonPath()
             .getList(".", FavoriteResponse.class);
         //then
         즐겨찾기_목록_조회됨(getFavoriteResponses);
 
         //when
-        ExtractableResponse<Response> deleteResponse = 즐겨찾기_삭제_요청(accessToken,
+        ExtractableResponse<Response> deleteResponse = FavoriteFixture.즐겨찾기_삭제_요청(accessToken,
             getURI(createResponse));
         //then
         즐겨찾기_삭제됨(deleteResponse);
@@ -146,13 +108,13 @@ public class FavoriteAcceptanceTest extends AcceptanceTest {
     @Test
     void delete_others_favorite() {
         //when
-        ExtractableResponse<Response> createResponse = 즐겨찾기_생성(accessToken, 강남역.getId(),
+        ExtractableResponse<Response> createResponse = FavoriteFixture.즐겨찾기_생성(accessToken, 강남역.getId(),
             정자역.getId());
         //then
         즐겨찾기_생성됨(createResponse);
 
         //when
-        ExtractableResponse<Response> deleteResponse = 즐겨찾기_삭제_요청(newAccessToken,
+        ExtractableResponse<Response> deleteResponse = FavoriteFixture.즐겨찾기_삭제_요청(newAccessToken,
             getURI(createResponse));
         //then
         즐겨찾기_삭제_에러_발생(deleteResponse);
