+++ src/test/java/nextstep/subway/favorite/FavoriteFixture.java
@@ -1,0 +1,45 @@
+package nextstep.subway.favorite;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.favorite.dto.FavoriteRequest;
+import org.springframework.http.MediaType;
+
+public class FavoriteFixture {
+
+    public static ExtractableResponse<Response> 즐겨찾기_생성(final String accessToken, Long source,
+        Long target) {
+        FavoriteRequest favoriteRequest = new FavoriteRequest(source, target);
+
+        return RestAssured
+            .given().log().all()
+            .auth().oauth2(accessToken)
+            .body(favoriteRequest)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().post("/favorites")
+            .then().log().all()
+            .extract();
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_목록_조회(final String accessToken) {
+        return RestAssured
+            .given().log().all()
+            .auth().oauth2(accessToken)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().get("/favorites")
+            .then().log().all()
+            .extract();
+    }
+
+    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(final String accessToken,
+        final String uri) {
+        return RestAssured
+            .given().log().all()
+            .auth().oauth2(accessToken)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().delete(uri)
+            .then().log().all()
+            .extract();
+    }
+}
